fix(qa): move browser-driver profile to src/qa/ (src/agents/*.ts must default-export createAgent)
flue auto-discovers every src/agents/*.ts as a top-level agent requiring a default createAgent(); the
browser-driver is a subagent PROFILE (defineAgentProfile) passed to the lead's subagents[]. In src/agents/
it built fine but CRASHED the prebuilt server at boot ('must default-export createAgent'). Moved to
src/qa/browser-driver.ts + fixed imports. Verified by booting dist/server.mjs (agent discovery now passes).
